// src/proto/models.ts
export abstract class ANConverter {
  importer: AppleNotesImporter;

  static protobufType: string;

  constructor(importer: AppleNotesImporter) {
    this.importer = importer;
  }

  abstract format(): Promise<string>;
}

export type ANConverterType<T extends ANConverter> = {
  new (importer: AppleNotesImporter, x: any): T;
  protobufType: string;
};

// SQLite types

export type SQLiteTagSpawned = {
  get(...query: any[]): Promise<SQLiteRow>;
  all(...query: any[]): Promise<SQLiteTable>;
  close(): void;
};

type SQLiteTable = SQLiteRow[];

interface SQLiteRow extends Record<string, any> {
  [member: string]: any;
}

// A few typings for internal use

export type ANAccount = {
  name: string;
  uuid: string;
  path: string;
};
// ...
export enum ANFolderType {
  Default = 0,
  Trash = 1,
  Smart = 3,
}

export interface ANAttachmentInfo extends Message {
  attachmentIdentifier: string;
  typeUti: string | ANAttachment;
}

export enum ANAttachment {
  Drawing = "com.apple.paper",
  DrawingLegacy = "com.apple.drawing",
  DrawingLegacy2 = "com.apple.drawing.2",
  Hashtag = "com.apple.notes.inlinetextattachment.hashtag",
  Mention = "com.apple.notes.inlinetextattachment.mention",
  InternalLink = "com.apple.notes.inlinetextattachment.link",
  PaperDocScan = "com.apple.paper.doc.scan",
  PaperDocPDF = "com.apple.paper.doc.pdf",
  Scan = "com.apple.notes.gallery",
  Table = "com.apple.notes.table",
  UrlCard = "public.url",
}